Extract file upload into reusable component
